Moved some functions from io to atomicUtils and removed some unused functions.
